% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOC_ph23_par.R
\name{getOC_ph23_par}
\alias{getOC_ph23_par}
\title{Get operating characteristics via simulations for the phase 2/3 design}
\usage{
getOC_ph23_par(
  ncore = 2,
  seed = 2024,
  nsim = 10,
  test.method = "dunnett",
  approach = "disjoint",
  n1 = c(50, 50, 50, 50),
  n2 = c(150, 150),
  duration1 = 8,
  duration2 = 12,
  enrollment_hold = 5,
  hazard = c(0.1, 0.09, 0.08, 0.07),
  orr = c(0.2, 0.25, 0.3, 0.4),
  rho = 0.7,
  dropout = c(0, 0, 0, 0),
  a1 = 1,
  a2 = 1,
  min_followup = 4,
  targetEventsIA_all = 240,
  targetEventsFA_all = 320,
  w = NULL,
  bound_z = c(2.44, 2),
  alpha = 0.025,
  update_bound = TRUE,
  nonselected_max_followup = NULL,
  dose_selection_endpoint = "ORR"
)
}
\arguments{
\item{ncore}{number of cores for parallel computing.}

\item{seed}{random seed for reproducibility.}

\item{nsim}{number of replicates}

\item{test.method}{method for getting adjusted p-value: \code{"dunnett"} or \code{"simes"}.}

\item{approach}{approach for combining data:
\code{"disjoint"}: disjoint-subjects approach;
\code{"ii"}: independent incremental approach.}

\item{n1}{total number of subjects to be enrolled in stage 1 for control and treatment arms, where
the number of treatment arms can be greater than 2.}

\item{n2}{total number of subjects to be enrolled in stage 2 for control and the selected treatment arm.}

\item{duration1}{enrollment duration in months for stage 1; must be an integer.}

\item{duration2}{enrollment duration in months for stage 2; must be an integer.}

\item{enrollment_hold}{holding period in month after stage 1 enrollment prior to enrollment of stage 2 patients.
0 means seamless enrollment.}

\item{hazard}{hazard rate for control and treatment arms.}

\item{orr}{ORR for control and treatment arms.}

\item{rho}{the correlation between ORR and survival based on Gaussian copula.}

\item{dropout}{dropout hazard rate for control and treatment arms.}

\item{a1}{weight parameter in cumulative enrollment pattern for stage 1. When \code{a=1}, it is uniform.}

\item{a2}{weight parameter in cumulative enrollment pattern for stage 2. When \code{a=1}, it is uniform.}

\item{min_followup}{minimum follow-up in month for stage 1 before dose selection. It should be smaller than
\code{enrollment_hold}. For example, a reasonable choice is \code{min_followup = 4} with \code{enrollment_hold=5}
which gives us at least 4 months follow-up for stage 1 patients and 1 month for dose selection.}

\item{targetEventsIA_all}{target number of events at IA for all subjects from both stage 1 and 2.}

\item{targetEventsFA_all}{target number of events at FA for all subjects from both stage 1 and 2.}

\item{w}{weights parameter for stage 1 data: for independent incremental approach, it will be length of 1;
for the disjoint-subjects approach, it will be length of 2 for IA and FA.}

\item{bound_z}{z-scale rejection boundaries and IA and FA.}

\item{alpha}{Type I error, always one-sided.}

\item{update_bound}{whether to re-calculate the FA rejection boundary using the updated
correlation between IA and FA test statistics.}

\item{nonselected_max_followup}{The final data cutoff time for non-selected arms.
The default \code{NULL} means the non-selected arms will not be followed up after dose selection.}

\item{dose_selection_endpoint}{either "ORR" or "Survival"}
}
\value{
It returns a matrix with each row corresponding to the analysis results for each trial,
where the analysis results include: "Selected dose", "IA time", "IA reject", "FA reject", "FA time",
"Study duration", "Total sample size", "observed correlation", and "actual FA boundary".
}
\description{
Get operating characteristics via simulations for the phase 2/3 design
}
\examples{
res <- getOC_ph23_par(ncore = 5, nsim=10)
apply(res, 2, mean, na.rm=TRUE)
}
