% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZ1.R
\name{getZ1}
\alias{getZ1}
\title{Get test statistic value using independent incremental approach}
\usage{
getZ1(
  dat,
  IAd_time,
  w = NULL,
  selected = 1,
  targetEvents,
  test.method = "dunnett"
)
}
\arguments{
\item{dat}{A time-to-event dataset returned from \link{sim_ph23}.}

\item{IAd_time}{the data cutoff time for dose selection.}

\item{w}{weight parameter for stage 1 data.}

\item{selected}{The selected dose level.}

\item{targetEvents}{The target number of events from both stages at which analysis is to be made.}

\item{test.method}{The method for getting adjusted p-value: \code{"dunnett"} or \code{"simes"}.}
}
\value{
It returns a list including \code{z} the test statistic, \code{obsEvents_stage1} the
observed number of events from stage 1 and \code{cut_time} the data cut time.
}
\description{
Get test statistic value using independent incremental approach
}
\examples{
d <- sim_ph23(n1 = c(20, 20, 20, 20), n2 = c(30, 30), duration1 = 4, enrollment_hold = 5)
getZ1(d, IAd_time = 13, selected = 2, targetEvents = 30, test.method="dunnett")
}
