% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ph23.R
\name{sim_ph23}
\alias{sim_ph23}
\title{Simulate a seamless phase 2/3 trial}
\usage{
sim_ph23(
  n1 = c(50, 50, 50, 50),
  n2 = c(150, 150),
  duration1 = 8,
  duration2 = 12,
  enrollment_hold = 5,
  hazard = c(0.1, 0.09, 0.08, 0.07),
  orr = c(0.2, 0.25, 0.3, 0.4),
  rho = 0.7,
  dropout = c(0, 0, 0, 0),
  a1 = 1,
  a2 = 1
)
}
\arguments{
\item{n1}{total number of subjects to be enrolled in stage 1 for control and treatment arms, where
the number of treatment arms can be greater than 2.}

\item{n2}{total number of subjects to be enrolled in stage 2 for control and the selected treatment arm.}

\item{duration1}{enrollment duration in months for stage 1; must be an integer.}

\item{duration2}{enrollment duration in months for stage 2; must be an integer.}

\item{enrollment_hold}{holding period in month after stage 1 enrollment prior to enrollment of stage 2 patients.
0 means seamless enrollment.}

\item{hazard}{hazard rate for control and treatment arms.}

\item{orr}{ORR for control and treatment arms.}

\item{rho}{the correlation between ORR and survival based on Gaussian copula.}

\item{dropout}{dropout hazard rate for control and treatment arms.}

\item{a1}{weight parameter in cumulative enrollment pattern for stage 1. When \code{a=1}, it is uniform.}

\item{a2}{weight parameter in cumulative enrollment pattern for stage 2. When \code{a=1}, it is uniform.}
}
\value{
A time-to-event dataset consists of
\itemize{
\item \code{stage}: 1=phase II; 2=phase III.
\item \code{trt}: 0=control; 1=dose 1; 2=dose 2; ...
\item \code{enterTime}: subject entry time relative to the study starting date
\item \code{response}: 0=no response; 1=response
\item \code{survTime}: observed event time
\item \code{event}: 0=censored; 1=event
\item \code{calendarTime}: total time on study, i.e. \code{enterTime} + \code{survTime}.
}
}
\description{
This function will simulate a phase 2/3 trial where patients will be randomized to all arms in stage 1
and only the selected arm and control arm will go into stage 2. Since we don't know the arm selection in
advance, we will simulated patients for all arms in stage 2 but with adjusted accrual rate assuming stage 2
only contains the selected arm and the control.
}
\examples{
d <- sim_ph23(n1 = c(50, 50, 50, 50), n2 = c(150, 150))
head(d)
}
