% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_single_arm.R
\name{sim_single_arm}
\alias{sim_single_arm}
\title{Simulate a single arm survival data}
\usage{
sim_single_arm(
  n = 50,
  duration = 2,
  hazard = 0.1,
  orr = 0.2,
  rho = 0.7,
  dropout = 0,
  a = 1
)
}
\arguments{
\item{n}{total number of subjects to be enrolled.}

\item{duration}{enrollment duration in months; must be an integer.}

\item{hazard}{hazard rate}

\item{orr}{objective response rate}

\item{rho}{the correlation between ORR and survival based on Gaussian copula.}

\item{dropout}{dropout hazard rate.}

\item{a}{weight parameter in cumulative enrollment pattern. When \code{a=1}, it is uniform.}
}
\value{
A time-to-event dataset consists of
\itemize{
\item \code{enterTime}: subject entry time relative to the study starting date
\item \code{response}: 0=no response; 1=response
\item \code{survTime}: observed event time
\item \code{event}: 0=censored; 1=event
\item \code{calendarTime}: total time on study, i.e. \code{enterTime} + \code{survTime}.
}
}
\description{
This function will simulate a single arm survival data with flexible enrollment distribution.
}
\examples{
d <- sim_single_arm(n=50)
head(d)
}
